content/svg/content/test/test_SVGAnimatedImageSMILDisabled.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=623945
-->
<head>
  <title>Test for Bug 623945</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=623945">Mozilla Bug 623945</a>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 623945 **/
/* This test makes sure we don't crash when using an animated SVG image with
 * the 'svg.smil.enabled' pref turned off. */

SimpleTest.waitForExplicitFinish();

function setSMILEnabled(enabled) {
  netscape.security.PrivilegeManager.enablePrivilege('UniversalXPConnect');
  var prefs = Components.classes["@mozilla.org/preferences-service;1"]
                 .getService(Components.interfaces.nsIPrefBranch);
  prefs.setBoolPref("svg.smil.enabled", enabled);
}

function iframeLoaded() {
  // Woo-hoo, we didn't crash! Declare success!
  ok(true, "got through the test without crashing");

  // Re-enable SMIL pref for future tests.
  setSMILEnabled(true);
  SimpleTest.finish();
}

function main() {
  // First, we turn off the SMIL pref:
  setSMILEnabled(false);

  // Then, we load an HTML document with an animated SVG image.
  var iframe = document.createElement("iframe");
  iframe.onload = iframeLoaded;
  iframe.src = "animated-svg-image-helper.html";
  document.getElementById("content").appendChild(iframe);
}

main();

</script>
</pre>
</body>
</html>
